requirements.txt
Django>=4.0,<5.0
mysqlclient
python-decouple
django-cors-headers
djangorestframework==3.15.2